+++ lib/thawani_payment.dart
@@ -1,5 +1,3 @@
 library thawani_payment;
-
-export 'widgets/pay_button.dart';
 export 'pay.dart';
 export 'viewmodel/thawani_customerdelete.dart';